polish repl and step
