+++ cmd/repl/repl_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package repl
+package main
 
 import (
 	"bufio"
@@ -33,6 +33,7 @@ import (
 	"sqlflow.org/sqlflow/pkg/sql"
 	"sqlflow.org/sqlflow/pkg/sql/codegen/attribute"
 	"sqlflow.org/sqlflow/pkg/sql/testdata"
+	"sqlflow.org/sqlflow/pkg/step"
 )
 
 var space = regexp.MustCompile(`\s+`)
@@ -60,31 +61,31 @@ func TestRunStmt(t *testing.T) {
 	os.Setenv("SQLFLOW_log_dir", "/tmp/")
 	session.DbConnStr = dbConnStr
 	currentDB = ""
-	output, err := GetStdout(func() error { return runStmt("show tables", true, "", dbConnStr) })
+	output, err := step.GetStdout(func() error { return runStmt("show tables", true, "", dbConnStr) })
 	a.Nil(err)
 	a.Contains(output, "Error 1046: No database selected")
 
-	output, err = GetStdout(func() error { return runStmt("use iris", true, "", dbConnStr) })
+	output, err = step.GetStdout(func() error { return runStmt("use iris", true, "", dbConnStr) })
 	a.Nil(err)
 	a.Contains(output, "Database changed to iris")
 
-	output, err = GetStdout(func() error { return runStmt("show tables", true, "", dbConnStr) })
+	output, err = step.GetStdout(func() error { return runStmt("show tables", true, "", dbConnStr) })
 	a.Nil(err)
 	a.Contains(output, "| TABLES IN IRIS |")
 
-	output, err = GetStdout(func() error {
+	output, err = step.GetStdout(func() error {
 		return runStmt("select * from train to train DNNClassifier WITH model.hidden_units=[10,10], model.n_classes=3, validation.select=\"select * from test\" label class INTO sqlflow_models.repl_dnn_model;", true, "", dbConnStr)
 	})
 	a.Nil(err)
 	a.Contains(output, "'global_step': 110")
 
-	output, err = GetStdout(func() error {
+	output, err = step.GetStdout(func() error {
 		return runStmt("select * from train to train xgboost.gbtree WITH objective=reg:squarederror, validation.select=\"select * from test\" label class INTO sqlflow_models.repl_xgb_model;", true, "", dbConnStr)
 	})
 	a.Nil(err)
 	a.Contains(output, "Evaluation result: ")
 
-	output, err = GetStdout(func() error {
+	output, err = step.GetStdout(func() error {
 		return runStmt("select * from train to explain sqlflow_models.repl_xgb_model;", true, "", dbConnStr)
 	})
 	a.Nil(err)
@@ -108,7 +109,7 @@ INTO sqlflow_models.repl_dnn_model;
 use sqlflow_models;
 show tables`
 	scanner := bufio.NewScanner(strings.NewReader(sql))
-	output, err := GetStdout(func() error { repl(scanner, "", dbConnStr); return nil })
+	output, err := step.GetStdout(func() error { repl(scanner, "", dbConnStr); return nil })
 	a.Nil(err)
 	a.Contains(output, "Database changed to iris")
 	a.Contains(output, `
@@ -136,7 +137,7 @@ func TestMain(t *testing.T) {
 	a := assert.New(t)
 	a.Nil(prepareTestDataOrSkip(t))
 	os.Args = append(os.Args, "-datasource", dbConnStr, "-e", "use iris; show tables", "-model_dir", "/tmp/repl_test")
-	output, _ := GetStdout(func() error { main(); return nil })
+	output, _ := step.GetStdout(func() error { main(); return nil })
 	a.Contains(output, `
 +----------------+
 | TABLES IN IRIS |
@@ -592,13 +593,6 @@ func TestTerminalCheck(t *testing.T) {
 	_, err := exec.LookPath("it2check")
 	a.Nil(err)
 	a.False(it2Check)
-
-	a.True(isHTMLSnippet("<div></div>"))
-	_, err = getBase64EncodedImage("")
-	a.Error(err)
-	image, err := getBase64EncodedImage(testImageHTML)
-	a.Nil(err)
-	a.Nil(imageCat(image)) // sixel mode
 }
 
 func TestGetTerminalColumnSize(t *testing.T) {
