+++ pkg/step/index.html
@@ -1,0 +1,16 @@
+<!-- Please don't delete this file!
+
+     For the reason, please refer to /index.html -->
+<html>
+<head>
+<meta name="go-import" content="sqlflow.org/sqlflow git https://github.com/sql-machine-learning/sqlflow">
+<meta http-equiv="refresh" content="0; url=https://github.com/sql-machine-learning/sqlflow" />
+<link rel="canonical" href="https://github.com/sql-machine-learning/sqlflow" />
+<script>
+	window.location.replace("https:\/\/github.com\/sql-machine-learning\/sqlflow");
+</script>
+</head>
+<body>
+	<h1>Redirecting to <a href="https://github.com/sql-machine-learning/sqlflow">https://github.com/sql-machine-learning/sqlflow</a></h1>
+</body>
+</html>
